pipeline refactor

+++ src/main/groovy/com/github/winplay02/gitcraft/mappings/yarn/FabricIntermediaryMappings.java
@@ -3,14 +3,17 @@ package com.github.winplay02.gitcraft.mappings.yarn;
 import com.github.winplay02.gitcraft.GitCraft;
 import com.github.winplay02.gitcraft.GitCraftConfig;
 import com.github.winplay02.gitcraft.mappings.Mapping;
-import com.github.winplay02.gitcraft.pipeline.Step;
+import com.github.winplay02.gitcraft.pipeline.MinecraftJar;
+import com.github.winplay02.gitcraft.pipeline.StepStatus;
 import com.github.winplay02.gitcraft.types.OrderedVersion;
 import com.github.winplay02.gitcraft.util.GitCraftPaths;
 import com.github.winplay02.gitcraft.util.RemoteHelper;
 import net.fabricmc.loom.api.mappings.layered.MappingsNamespace;
+import net.fabricmc.mappingio.MappingVisitor;
 import net.fabricmc.mappingio.MappingWriter;
 import net.fabricmc.mappingio.format.MappingFormat;
 import net.fabricmc.mappingio.format.tiny.Tiny1FileReader;
+import net.fabricmc.mappingio.format.tiny.Tiny2FileReader;
 import net.fabricmc.mappingio.tree.MemoryMappingTree;
 
 import java.io.BufferedReader;
@@ -38,27 +41,43 @@ public class FabricIntermediaryMappings extends Mapping {
 		return mcVersion.compareTo(GitCraft.config.manifestSource.getMetadataProvider().getVersionByVersionID(GitCraftConfig.FABRIC_INTERMEDIARY_MAPPINGS_START_VERSION_ID)) >= 0;
 	}
 
+	@Override
+	public boolean doMappingsExist(OrderedVersion mcVersion, MinecraftJar minecraftJar) {
+		// fabric intermediary is provided for the merged jar
+		return minecraftJar == MinecraftJar.MERGED && doMappingsExist(mcVersion);
+	}
+
+	@Override
+	public boolean canMappingsBeUsedOn(OrderedVersion mcVersion, MinecraftJar minecraftJar) {
+		// the merged mappings can be used for all jars
+		return true;
+	}
+
 	protected static String mappingsIntermediaryPathQuirkVersion(String version) {
 		return GitCraftConfig.yarnInconsistentVersionNaming.getOrDefault(version, version);
 	}
 
 	@Override
-	public Step.StepResult prepareMappings(OrderedVersion mcVersion) throws IOException {
-		Path mappingsFile = getMappingsPathInternal(mcVersion);
+	public StepStatus provideMappings(OrderedVersion mcVersion, MinecraftJar minecraftJar) throws IOException {
+		// fabric intermediary is provided for the merged jar
+		if (minecraftJar != MinecraftJar.MERGED) {
+			return StepStatus.NOT_RUN;
+		}
+		Path mappingsFile = getMappingsPathInternal(mcVersion, minecraftJar);
 		if (Files.exists(mappingsFile) && validateMappings(mappingsFile)) {
-			return Step.StepResult.UP_TO_DATE;
+			return StepStatus.UP_TO_DATE;
 		}
 		Files.deleteIfExists(mappingsFile);
 		Path mappingsV1 = getMappingsPathInternalV1(mcVersion);
-		Step.StepResult downloadResult = RemoteHelper.downloadToFileWithChecksumIfNotExistsNoRetryGitHub("FabricMC/intermediary", "master", String.format("mappings/%s.tiny", mappingsIntermediaryPathQuirkVersion(mcVersion.launcherFriendlyVersionName())), new RemoteHelper.LocalFileInfo(mappingsV1, null, "intermediary mapping", mcVersion.launcherFriendlyVersionName()));
+		StepStatus downloadStatus = RemoteHelper.downloadToFileWithChecksumIfNotExistsNoRetryGitHub("FabricMC/intermediary", "master", String.format("mappings/%s.tiny", mappingsIntermediaryPathQuirkVersion(mcVersion.launcherFriendlyVersionName())), new RemoteHelper.LocalFileInfo(mappingsV1, null, "intermediary mapping", mcVersion.launcherFriendlyVersionName()));
 		MemoryMappingTree mappingTree = new MemoryMappingTree();
-		try (BufferedReader clientBufferedReader = Files.newBufferedReader(mappingsV1, StandardCharsets.UTF_8)) {
-			Tiny1FileReader.read(clientBufferedReader, mappingTree);
+		try (BufferedReader br = Files.newBufferedReader(mappingsV1, StandardCharsets.UTF_8)) {
+			Tiny1FileReader.read(br, mappingTree);
 		}
-		try (MappingWriter w = MappingWriter.create(mappingsFile, MappingFormat.TINY_2_FILE)) {
-			mappingTree.accept(w);
+		try (MappingWriter writer = MappingWriter.create(mappingsFile, MappingFormat.TINY_2_FILE)) {
+			mappingTree.accept(writer);
 		}
-		return Step.StepResult.merge(downloadResult, Step.StepResult.SUCCESS);
+		return StepStatus.merge(downloadStatus, StepStatus.SUCCESS);
 	}
 
 	protected Path getMappingsPathInternalV1(OrderedVersion mcVersion) {
@@ -66,7 +85,15 @@ public class FabricIntermediaryMappings extends Mapping {
 	}
 
 	@Override
-	protected Path getMappingsPathInternal(OrderedVersion mcVersion) {
+	protected Path getMappingsPathInternal(OrderedVersion mcVersion, MinecraftJar minecraftJar) {
 		return GitCraftPaths.MAPPINGS.resolve(mcVersion.launcherFriendlyVersionName() + "-intermediary.tiny");
+	}
+
+	@Override
+	public void visit(OrderedVersion mcVersion, MinecraftJar minecraftJar, MappingVisitor visitor) throws IOException {
+		Path path = getMappingsPathInternal(mcVersion, MinecraftJar.MERGED);
+		try (BufferedReader br = Files.newBufferedReader(path)) {
+			Tiny2FileReader.read(br, visitor);
+		}
 	}
 }
